+++ src/neural/backends/cuda/layers.cc
@@ -32,9 +32,9 @@
 
 #include "cuda_common.h"
 #include "kernels.h"
+#include "neural/backends/shared/activation.h"
 #include "neural/network.h"
-#include "neural/shared/activation.h"
-#include "neural/shared/attention_policy_map.h"
+#include "neural/tables/attention_policy_map.h"
 #include "utils/fp16_utils.h"
 
 namespace lczero {
