+++ src/neural/backends/cuda/network_cudnn.cc
@@ -35,11 +35,11 @@
 #include "inputs_outputs.h"
 #include "kernels.h"
 #include "layers.h"
+#include "neural/backends/shared/activation.h"
 #include "neural/factory.h"
 #include "neural/network_legacy.h"
-#include "neural/shared/activation.h"
-#include "neural/shared/attention_policy_map.h"
-#include "neural/shared/policy_map.h"
+#include "neural/tables/attention_policy_map.h"
+#include "neural/tables/policy_map.h"
 #include "utils/bititer.h"
 #include "utils/exception.h"
