+++ src/neural/backends/onednn/network_onednn.cc
@@ -35,8 +35,8 @@
 #include "layers.h"
 #include "neural/factory.h"
 #include "neural/network_legacy.h"
-#include "neural/shared/attention_policy_map.h"
-#include "neural/shared/policy_map.h"
+#include "neural/tables/attention_policy_map.h"
+#include "neural/tables/policy_map.h"
 #include "utils/bititer.h"
 #include "utils/exception.h"
