+++ src/neural/backends/opencl/network_opencl.cc
@@ -24,12 +24,12 @@
 
 #include "neural/backends/opencl/OpenCL.h"
 #include "neural/backends/opencl/OpenCLParams.h"
+#include "neural/backends/shared/activation.h"
+#include "neural/backends/shared/winograd_filter.h"
 #include "neural/factory.h"
 #include "neural/network.h"
 #include "neural/network_legacy.h"
-#include "neural/shared/activation.h"
-#include "neural/shared/policy_map.h"
-#include "neural/shared/winograd_filter.h"
+#include "neural/tables/policy_map.h"
 #include "utils/bititer.h"
 #include "utils/exception.h"
 #include "utils/filesystem.h"
