+++ src/neural/onnx/converter.cc
@@ -38,9 +38,9 @@
 #include "neural/network_legacy.h"
 #include "neural/onnx/adapters.h"
 #include "neural/onnx/builder.h"
-#include "neural/shared/activation.h"
-#include "neural/shared/attention_policy_map.h"
-#include "neural/shared/policy_map.h"
+#include "neural/tables/activation_function.h"
+#include "neural/tables/attention_policy_map.h"
+#include "neural/tables/policy_map.h"
 #include "proto/net.pb.h"
 #include "utils/bf16_utils.h"
 #include "utils/exception.h"
