+++ src/neural/tables/activation_function.h
@@ -18,9 +18,6 @@
 
 #pragma once
 
-#include <cstddef>
-#include <vector>
-
 namespace lczero {
 // The following list matches the one in net.proto. Ideally this would be done
 // by including proto/net.pb.h, but this is incompatible with nvcc.
@@ -36,31 +33,5 @@ enum ActivationFunction {
   ACTIVATION_RELU_2 = 8,
   ACTIVATION_SOFTMAX = 9,
 };
-
-struct Activations {
-    ActivationFunction default_activation = ACTIVATION_RELU;
-    ActivationFunction smolgen_activation = ACTIVATION_SWISH;
-    ActivationFunction ffn_activation = ACTIVATION_RELU_2;
-};
-
-// Softmax activation
-void SoftmaxActivation(const size_t size, const float* input, float* output);
-
-void BiasResidual(const size_t batch_size, const size_t channels, float* data,
-                  const float* biases, const float* eltwise,
-                  const ActivationFunction activation);
-
-void BiasActivate(const size_t batch_size, const size_t channels, float* data,
-                  const float* biases,
-                  const ActivationFunction activation);
-
-float Activate(const float val, const ActivationFunction activation);
-
-void Activate(const size_t len, const float* data, const float* bias,
-              float* output, const ActivationFunction activation);
-
-void Activate(const size_t len, float gamma, const float* data,
-              const float* bias, float beta, float* out,
-              const ActivationFunction activation);
 
 }  // namespace lczero
